wip improving test suite and fixing tests
